Update expectation filenames to match snake_case of their defined Expectations
