+++ contrib/capitalone_dataprofiler_expectations/capitalone_dataprofiler_expectations/expectations/__init__.py
@@ -1,8 +1,8 @@
 # Make sure to include any Expectations your want exported below!
 
-from .expect_column_values_are_probabilistically_greater_than_or_equal_to_threshold import (
+from .expect_column_values_to_be_greater_than_or_equal_to_threshold import (
     ExpectColumnValuesToBeGreaterThanOrEqualToThreshold,
 )
-from .expect_column_values_confidence_to_be_greater_than_or_equal_to_threshold_for_data_label import (
-    ExpectColumnsValuesConfidenceForDataLabelToBeGreaterThanOrEqualtoThreshold,
+from .expect_column_values_confidence_for_data_label_to_be_greater_than_or_equal_to_threshold import (
+    ExpectColumnValuesConfidenceForDataLabelToBeGreaterThanOrEqualtoThreshold,
 )
