+++ contrib/capitalone_dataprofiler_expectations/capitalone_dataprofiler_expectations/expectations/expect_column_values_confidence_for_data_label_to_be_greater_than_or_equal_to_threshold.py
@@ -27,42 +27,66 @@ from great_expectations.expectations.metrics import (
 )
 
 
-class ColumnValuesConfidenceToBeGreaterThanOrEqualToThreshold(ColumnMapMetricProvider):
-    """MetricProvider Class for determining if the dataprofiler's model confidence value is\
-        greater than or equal to the user-specified threshold"""
+class ColumnValuesConfidenceForDataLabelToBeGreaterThanOrEqualToThreshold(
+    ColumnMapMetricProvider
+):
+    """MetricProvider Class for Data Label Probability greater than \
+    or equal to the user-specified threshold"""
 
     # This is the id string that will be used to reference your metric.
-    condition_metric_name = (
-        "column_values.prediction_confidence_greater_than_or_equal_to_threshold"
-    )
+    condition_metric_name = "column_values.prediction_confidence_for_data_label_greater_than_or_equal_to_threshold"
 
-    condition_value_keys = ("threshold",)
+    condition_value_keys = (
+        "threshold",
+        "data_label",
+    )
 
     # This method implements the core logic for the PandasExecutionEngine
     @column_condition_partial(engine=PandasExecutionEngine)
-    def _pandas(cls: Any, column: str, threshold: float, **kwargs: Any) -> np.ndarray:
+    def _pandas(
+        cls: Any, column: str, threshold: float, data_label: str, **kwargs: Any
+    ) -> np.ndarray:
         """
         Implement the yes/no question for the expectation
         """
         labeler = dp.DataLabeler(labeler_type="structured")
         labeler.postprocessor.set_params(is_pred_labels=False)
-        results = labeler.predict(column, predict_options={"show_confidences": True})
-        return (
-            np.choose(results["pred"].astype(int, copy=False), results["conf"].T)
-            >= threshold
+
+        results = labeler.predict(
+            column,
+            predict_options={"show_confidences": True},
         )
 
+        if data_label.upper() in labeler.label_mapping.keys():
+            data_label_ind = labeler.label_mapping[data_label.upper()]
+        else:
+            raise ValueError(
+                """
+                The only values acceptable for the data label parameter are as follows:
+                ['PAD', 'UNKNOWN', 'ADDRESS', 'BAN', 'CREDIT_CARD', 'DATE', 'TIME', 'DATETIME',\
+                    'DRIVERS_LICENSE', 'EMAIL_ADDRESS', 'UUID', 'HASH_OR_KEY', 'IPV4', 'IPV6',\
+                    'MAC_ADDRESS', 'PERSON', 'PHONE_NUMBER', 'SSN', 'URL', 'US_STATE', 'INTEGER',\
+                    'FLOAT', 'QUANTITY', 'ORDINAL']
+            """
+            )
+        data_label_conf = results["conf"][:, data_label_ind]
+        return data_label_conf >= threshold
 
-class ExpectColumnValuesToBeGreaterThanOrEqualToThreshold(ColumnMapExpectation):
+
+class ExpectColumnValuesConfidenceForDataLabelToBeGreaterThanOrEqualtoThreshold(
+    ColumnMapExpectation
+):
     """
-    This function builds upon the custom column map expectations of Great Expectations. This function asks a yes/no question of each row in the user-specified column; namely, does the confidence threshold provided by the DataProfiler model exceed the user-specified threshold.
+    This function builds upon the custom column map expectations of Great Expectations. This function asks the question a yes/no question of each row in the user-specified column; namely, does the confidence threshold provided by the DataProfiler model exceed the user-specified threshold.
 
     Args:
         column (str): The column name that you want to check.
+        data_label(str): The data label for which you want to check confidences against the threshold value
         threshold (float): The value, usually as a decimal (e.g. .32), you want to use to flag low confidence predictions
 
-    df.expect_column_values_to_probabilistically_match_data_label(
+    df.expect_column_values_confidence_for_data_label_to_be_greater_than_or_equal_to_threshold(
         column,
+        data_label=<>,
         threshold=float(0<=1)
     )
     """
@@ -117,18 +141,16 @@ class ExpectColumnValuesToBeGreaterThanOrEqualToThreshold(ColumnMapExpectation):
                     "title": "positive_test_with_column_one",
                     "exact_match_out": False,
                     "include_in_gallery": True,
-                    "in": {"column": "OPEID6", "threshold": 0.00},
+                    "in": {"column": "ZIP", "data_label": "ADDRESS", "threshold": 0.00},
                     "out": {
                         "success": True,
-                        "unexpected_index_list": [],
-                        "unexpected_list": [],
                     },
                 },
                 {
-                    "title": "error_test_threshold",
+                    "title": "failing_test_with_column_one",
                     "exact_match_out": False,
                     "include_in_gallery": True,
-                    "in": {"column": "ZIP", "threshold": 0.61},
+                    "in": {"column": "ZIP", "data_label": "ADDRESS", "threshold": 1.00},
                     "out": {
                         "success": False,
                     },
@@ -139,19 +161,19 @@ class ExpectColumnValuesToBeGreaterThanOrEqualToThreshold(ColumnMapExpectation):
 
     # This is the id string of the Metric used by this Expectation.
     # For most Expectations, it will be the same as the `condition_metric_name` defined in your Metric class above.
-    map_metric = (
-        "column_values.prediction_confidence_greater_than_or_equal_to_threshold"
-    )
+    map_metric = "column_values.prediction_confidence_for_data_label_greater_than_or_equal_to_threshold"
 
     # This is a list of parameter names that can affect whether the Expectation evaluates to True or False
     success_keys = (
         "threshold",
+        "data_label",
         "mostly",
     )
 
     # This dictionary contains default values for any parameters that should have default values
     default_kwarg_values = {
         "threshold": None,
+        "data_label": None,
         "result_format": "BASIC",
         "include_config": True,
         "catch_exceptions": False,
@@ -170,6 +192,6 @@ class ExpectColumnValuesToBeGreaterThanOrEqualToThreshold(ColumnMapExpectation):
 
 if __name__ == "__main__":
     diagnostics_report = (
-        ExpectColumnValuesToBeGreaterThanOrEqualToThreshold().run_diagnostics()
+        ExpectColumnValuesConfidenceForDataLabelToBeGreaterThanOrEqualtoThreshold().run_diagnostics()
     )
     print(diagnostics_report.generate_checklist())
